requirements.txt
streamlit==1.32.2
pikepdf==8.11.2
Pillow==10.2.0 